2 этап(iter 62)

+++ main-service/src/main/java/ru/practicum/event/dto/EventDtoOut.java
@@ -31,5 +31,5 @@ public class EventDtoOut {
     String state;
     @NotBlank
     String title;
-    Long views;
+    Integer views;
 }
